requirements.txt
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
anyio==3.7.1
attrs==24.2.0
certifi==2024.12.14
cffi==1.15.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
config==0.5.1
crcmod==1.7
crefi==2.0.9
cryptography==44.0.0
exceptiongroup==1.2.2
Flask==2.2.5
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
importlib-metadata==6.7.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.5
jmespath==0.10.0
jsonschema==4.17.3
logger==1.4
MarkupSafe==2.1.5
oss2==2.19.1
pip==24.0
pkgutil_resolve_name==1.3.10
pycparser==2.21
pycryptodome==3.21.0
pyproject==1.3.1
pyproject-toml==0.0.12
pyrsistent==0.19.3
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
setuptools==68.0.0
six==1.17.0
sniffio==1.3.1
toml==0.10.2
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
wheel==0.42.0
zipp==3.15.0
